refactor(host): split Program.cs into composition extension methods (#187)
Extract the 830-line backend/src/B3.Trading.Host/Program.cs into focused
extension methods under B3.Trading.Host/Composition/ and move the inline
hosted services to B3.Trading.Host/Hosted/. Final Program.cs is 52 lines
of pure orchestration.

Composition extensions (one public static class per file):
- AddTradingAuth                 — JWT bearer, authz, user store, rate limiter
- AddTradingApplicationCore      — registries/books/processor/sinks/algo + drain
- AddTradingDataProtection       — DP keys persisted under PersistenceOptions
- AddTradingPersistence          — WAL/snapshot/recovery/seeds
- AddTradingRisk                 — risk options, checks, pipeline, accountants
                                   (also calls AddTradingMarketData since the
                                   risk pipeline is the only IReferencePrice
                                   consumer)
- AddTradingExchangeGateway      — Stub/Mock/Real/Unavailable gateway,
                                   ExchangeStatus, IFirmDirectory; preserves
                                   the early-vs-late ExchangeOptions read
                                   required for Real-mode hosted services
- MapTradingEndpoints            — every app.Map* call

Hosted services moved out of Program.cs (still internal):
- Hosted/EntryPointRouterStarter.cs
- Hosted/FirmGatewayConnector.cs

Post-Build helpers grouped into internal TradingHostStartup:
- RegisterMetricsSources
- ValidateBootGuards (auth signing key / ER injection / FIXP listener)
- RunRecoveryAndSeedingAsync

Behavior is unchanged. Service registration ORDER is preserved one-for-one
so options-binding-before-consumer and hosted-service-start-order semantics
match pre-#187. Internal types stay internal. No other layer touched. All
864 tests pass; dotnet format clean.

Closes #187
